fix: duplicate code
